Tạo luồng xử lý đơn hàng từ backend
